# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Interaction {
  id: number;
  height: number;
  figure: string;
  age: number;
  company: string;
  attitude: string;
  myMood: string;
  myPerformance: string;
  space: string;
  notes: string;
  lessonLearned: string;
  success: boolean;
  createdAt: string;
}

export interface InteractionInput {
  height: number;
  figure: string;
  age: number;
  company: string;
  attitude: string;
  myMood: string;
  myPerformance: string;
  space: string;
  notes: string;
  lessonLearned: string;
  success: boolean;
}

export interface InteractionStats {
  totalInteractions: number;
  /** @nullable */
  firstInteractionDate: string | null;
  /** @nullable */
  lastInteractionDate: string | null;
  totalDaysPassed: number;
  totalDaysActive: number;
  successRate: number;
}

export interface ChartDataPoint {
  label: string;
  interactions: number;
  successes: number;
  successRate: number;
}

export interface ChartData {
  granularity: string;
  points: ChartDataPoint[];
}