# lib/api-zod/src/generated/types/chartData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChartDataPoint } from './chartDataPoint';

export interface ChartData {
  granularity: string;
  points: ChartDataPoint[];
}