# lib/api-zod/src/generated/types/interaction.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Interaction {
  id: number;
  height: number;
  figure: string;
  age: number;
  company: string;
  attitude: string;
  myMood: string;
  myPerformance: string;
  space: string;
  notes: string;
  lessonLearned: string;
  success: boolean;
  createdAt: string;
}