# lib/api-zod/src/generated/types/interactionStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InteractionStats {
  totalInteractions: number;
  /** @nullable */
  firstInteractionDate: string | null;
  /** @nullable */
  lastInteractionDate: string | null;
  totalDaysPassed: number;
  totalDaysActive: number;
  successRate: number;
}